requirements.txt
pillow
numpy
h5py>=2.5.0
# planned:
#   blosc-hdf5-plugin
#   python-blosc
#   openPMD-api
